Make kernel driver detaching use old libusb api
The travis build system uses an older ubuntu
so the libusb there is missing the auto
detach function. The new api is only about
a year old. Using this older method should
help with porting to non-linux systems.

A minor downside of this is we no longer auto
re-attach the kernel driver. This is okay for
our use case so the functionality has not be
re-implemented on our end.