Introduced new "-P" option to supply a desired port
Before, we had only the choice of using a random free port (no option)
or select a port but error out if the port is already taken by
something else ("-p" option). Now we have also the "-P" option where
we also select a port but if the selected port is busy we search for a
free one incrementing the port number by 1 until getting
successful. This way we have our printer(s) usualkly on the same port
and it is eaier for the user to call the configuration web interface
of the printer.