Return process ID of the daemon process on terminal.
On tests it turns out that ippusbxd does not reliably terminate when
the printer is turned off or unplugged. Therefore we need a way to
easily kill the daemon externally. To make this easier we do not only
output the port on the console but also the process ID of the daemon
process.