Fixed bug in chunked transfers to the printer
Chunked HTTP transfers allow streaming and therefore they are used to
send larger amounts of data (Usually print jobs) to the printer. There
was a bug that made an HTTP message being considered completed when
the buffer space was full, ignoring the fact that packages can be
larger and that therefore the space gets expanded if necessary. This
made print jobs (except tiny ones, probably the ones smaller than the
initial buffer size of 4kB) hanging but instructions and queries
(which are transfered in content-length mode) working normally.