More verbose debug logging
Added new logs to the packet_check_completion() function in http.c

Add pointer addresses for message and packet to service_connection()
function in ippusbxd.c and to usb_conn_packet_send() function in
usb.c, to see whether log messages belong to the same thread.

Added log about the remaining bytes to be treated to the loop in the
tcp_packet_get() function in tcp.c.