Add no-printer debug/development mode
This new mode allows to run the daemon and to open the port on
localhost without having an IPP-over-USB printer. It simply ignores
the content of any request from a client and answers the request with
a short message in HTML, which especially means that if one accesses
with a browser, the message gets displayed.

This mode is for testing and debugging the TCP interface if one has no
IPP-over-USB printer handy, for example for security-auditing or for
regression testing during package build on a distribution's build
servers.

The mode is activated by the "-N" command line option and makes the
options for identifying a USB printer ignored. It does not make sense
for production use.