Simplified usage of select() for simultaneous listening via IPv4 and IPv6
On the introducion of select() to simultaneously listen on the
assigned port on localhost both via IPv4 and via IPv6 unnecessarily a
timeout was used and the select() call repeated in a loop if the
timeout expired without incoming data. Now the code is simplified to
not use a timeout and a loop but let select() wait for input
indefinitely.