Make the unplug callback actually work
To identify whether an unplug event is really from the printer which
we are actually connecting, we use the is_our_device() function which
is based on vendor ID, product ID, and serial number, but if we get an
unplug event from a device, the event does not supply the serial
number and is_our_device() polls the serial number from the device, in
which it does not succeed as the device is turned off and so the
unplug callback gets aborted. Now we note the bus and device addresses
on start-up of ippusbxd and is_our_device() falls back to using these
when the device is turned off, leading to a reliable shutdown of
ippusbxd when unplugging or turning off the printer.