Assign the packet type HTTP_HEADER_ONLY correctly to packets ehich are server's answer
The function packet_find_type() was only able to categorize packets
with GET requests as HTTP_HEADER_ONLY and not the server's answers
which simply acknowlege the submission of a command or of data to the
server. This made them categorized as unknown type and more data tried
to get read from the server when in reality there is no more data to
read.