Make sure that if libusb_bulk_transfer() times out, that no data gets lost or sent twice
On the USB side, data is exchanged with the printer using the
libusb_bulk_transfer() function. This function has a timeout facility
for the case that the device does not answer, but it can also time out
if a part of the data has already been sent or received. Now we take
this into account and if the function fails with a timeout we still
check the number of bytes transferred to not loose this data or send
data twice.