Let ippusbxd advertise the connected printer via Bonjour
To let software like ippfind or cups-browsed find an IPP-over-USB
printer we broadcast the printer which we are mapping to a network
interface via Bonjour (using Avahi).

As Avahi is not able to broadcast services on the loopback interface
(as the loopback interface does not support multicast), we allow
supplying any other interface on the command line, making the printer
socket be bound there and then the broadcast done on this interface.

To make the printer still appearing only on the local machine it is
recommended to use the dummy interface, setting it up via

sudo modprobe dummy

and one of

sudo ifconfig dummy0 10.0.0.1 netmask 255.255.255.0 multicast

or

sudo ifconfig dummy0 up multicast

After that one runs ippusbxd with the "-i dummy0" option and the
printer is available on the desired port on 10.0.0.1.