Take VID/PID as hex and bus/device as decimal numbers on the ippusbxd command line.
VID and PID are usually expressed in hexadecimal numbers and USB bus
and device usually in decimal but padded with zeros to three digits,
see also the output of the "lsusb" command.

Letting ippusbxd determine the base for the numbers supplied on the
command line automatically, bus and device are mis-interpreted as
octal numbers due to the leading zeros. Therefore we set fixed bases
now, hexaecimal for VID and PID and decimal for bus and device.

Note that even UDEV supplies bus and device with the leading zeros.