Added instructions and sample configuration files for installation
The readme.md file is updated to take into account the new dependency
on Avahi for broadcasting the presence of the printer on the local
system and instructions are added to set up fully automatic support
for IPP-over-USB printers on modern Linux systems so that print queues
for driverless (IPP Everywhere or AirPrint) printing can be easily set
up or even be set up and taken down fully automatically by
cups-browsed.

Sample UDEV rules and systemd service files are also added for quick
and easy setup.