Do not include the "adminurl = ..."  entry in the broadcasted TXT record
ippusbxd broadcasted a wrong URL for the admin web interface of the
printer (http://localhost:<port>/). As we do not have decided on a
solution, we temporarily do not broadcast the URL for the printer's
web interface.

For discovery and setup of the printer the "adminurl = ..." text
record entry is not needed and the URL of the web interface can also
be retrieved from the printer via IPP.