Renamed all occurences of "Bonjour" to "DNS-SD"
The facility for services being advertised by the machine providing
the service is called "DNS-SD" (Domain Name Service - Service
Discovery) and no "Bonjour". Bonjour is Apple's software to provide
this facility on Mac OS X, like Avahi is the software used for the
same purpose on Linux and other operating systems.