Unregister printer from DNS-SD when unplugging
Before, the unlug callback did nothing more than stopping ippusbxd
when the unplug event was identified with our printer. No DNS-SD
shutdown was performed and so the printer which is not present anymore
stayed registered. Now we do a DNS-SD shutdown so that the DNS-SD
entry goes away with the printer.