Added handler for SIGTERM and SIGINT for clean termination
To assure that the USB communication between the printer and the host
machine is cleanly terminated and the printer is unregistered from
DNS-SD when stopping the daemon via SIGTERM ("kill -TERM") or SIGINT
(Ctrl+C) a handler is added now which is called when one of these
signals is caught. The handler sets a global flag and then all loops
are terminated, making the daemon going through the regular exit
procedure after finishing the current packet.

Also done some corrections on debug messages, especially adding the
thread number.