When registering printer for DNS-SD broadcasting do it also for the "lo" interface
Current Avahi does not support the non-multicast-capable loopback
interfaces, so it will not advertise a service on a loopback interface
("lo", localhost") when one registers it. Therefore ippusbxd has
turned off DNS-SD support when the USB printer was mirrored onto the
"lo" interface.

But according to the PWG standard for IPP-over-USB the printer should
be mirrored on a free port on localhost and in addition advertised via
DNS-SD. This is only possible by modifying Avahi to support loopback
interfaces and registering the printer also if we use the "lo"
interface.

Avahi is not yet modified, but there is no fatal error caused by
registering a service on a loopback interface. ippusbxd carries on
after the registration and current Avahi simply does not advertise the
service. Therefore we drop the artificial restriction of turning of
DNS-SD support with the printer on a loopback interface. In case of
problems there is still the "-B" command line option to manually turn
off DNS-SD support.