Make ippusbxd run independent on whether avahi-daemon is running
To improve the stability of ippusbxd it is now independent of whether
avahi-daemon is actually running. It simply registers the printer when
avahi-daemon comes up and unregisters the printer when avahi-daemon
disappears.

This way it is no problem to restart avahi-daemon while ippusbxd is
running or to start ippusbxd before avahi-daemon gets started. It is
assured that the printer is always advertised as long as avahi-daemon
is running.