Rename global variable "sigterm" into "terminate"
In the beginning this variable was simply to indicate whether we got a
SIGTERM so that all threads can cleanly shut down. Now we use it also
for fatal errors to get a clean shutdown instead of a hard
"exit(1)". This way it got a general termination request flag and
should be named as such.