Reset printer on shutdown, improvements for thread termination
When we shut down the daemon via SIGINT or SIGTERM we reset the
printer now using the libusb_reset_device() function, so that if the
shutdown has occured while the printer was still in use all buffers
get cleared, pages ejected, ... to get the printer into a defined
state, ready for the next use.

Spinlock loops in the USB functions to get blocking behavior when
using non-blocking functions are now stopped when we shut down.

When a thread closes, we tell in the debug message whether it closed
due to the end of the communication thread or due to a daemon shutdown
request.