Check termination flag in all communication and wait loops
In most of the loops which run in the USB<->TCP communication threads
the termination flag g_options.termminate was checked, but not in
all. This could make the communication to carry on for a significant
time after a shutdown request and even to infinite loops sometimes.

Now all loops in functions for sending and receiving data and also the
loop to wait for a free USB interface check the termination flag and
abort immediately in case of shutdown.