Get thread handle correctly into the thread registration
USB<->TCP threads got registered before they get created and the
registration function copied the thread handle before it was
generated by the thread creation, leading to random data be
copied instead of a thread handle and so thread cancellation
on shutdown of ippusbxd not working.

Now we simply register a pointer to the thread's argument field
instead of copying individual arguments. This way the thread
handles get accessible through the registration list and so
threads get actually cancelled.