Improvements on USB unplug event handler
Let the unplug event handler not only shut down the DNS-SD advertising
but also the TCP socket connections to assure that address/port gets
cleanly unbound.

Use libusb_handle_events_timeout_completed() instead of
libusb_handle_events_completed() with a 1-second timeout in the thread
for observing USB events, to check the termination flag and close the
thread on ippusbxd shutdown.

Let the main thread wait for the thread to observe USB events to
finish before finally resetting the printer.