If we mirror our printer to a port of localhost, advertise its web interface
If we mirror our USB printer to a port on localhost (loopback
interface), we can easily DNS-SD-advertise the web admin interface of
the printer as "http://localhost:PORT/", whereas using other
interfaces causes problems with the host name.

Therefore we add the "adminurl" field to the TXT record of our DNS-SD
advertising only if we have mirrored the printer to the loopback
interface.