Added support for the page description language PCLm in the DNS-SD record
In addition to the page description languages PDF, PWG Raster, and
Apple Raster there is also HP's PCLm used for driverless printing
(Mopria and Wi-Fi Direct standards). As we have started to add support
for PCLm to cups-filters ippusbxd needs to advertise PCLm support on
appropriate printers in its DNS-SD records so that CUPS and
cups-filters (or any other client) can make use of it.